+++ vocaloid/utils/window.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <math.h>
 #include <vector>
+#include <stdint.h>
 using namespace std;
 namespace vocaloid {
 
@@ -17,7 +18,7 @@ namespace vocaloid {
 	* Calculates rectangle window coefficients.
 	*/
 	template <typename T>
-	void Rectangle(int N, T amp, vector<T> &win) {
+	void Rectangle(uint64_t N, T amp, vector<T> &win) {
 		for (int i = 0; i<(N + 1) / 2; ++i) {
 			win[i] = amp;
 			win[N - 1 - i] = win[i];
@@ -28,7 +29,7 @@ namespace vocaloid {
 	* Calculates bartlett window coefficients.
 	*/
 	template <typename T>
-	void Bartlett(int N, T amp, vector<T> &win) {
+	void Bartlett(uint64_t N, T amp, vector<T> &win) {
 		for (int i = 0; i<(N + 1) / 2; ++i) {
 			win[i] = amp * 2 * i / (N - 1);
 			win[N - 1 - i] = win[i];
@@ -39,7 +40,7 @@ namespace vocaloid {
 	* Calculates hanning window coefficients.
 	*/
 	template <typename T>
-	void Hanning(int N, vector<T> &win) {
+	void Hanning(uint64_t N, vector<T> &win) {
 		for (int i = 0; i < N; ++i) {
 			win[i] = 0.5f - 0.5f*cos(M_PI * 2 * i / N);
 		}
@@ -49,7 +50,7 @@ namespace vocaloid {
 	* Calculates hamming window coefficients.
 	*/
 	template <typename T>
-	void Hamming(int N, vector<T> &win) {
+	void Hamming(uint64_t N, vector<T> &win) {
 		for (int i = 0; i < N; ++i) {
 			win[i] = 0.54 - 0.46*cos(M_PI * 2 * i / (N - 1.0));
 		}
@@ -59,7 +60,7 @@ namespace vocaloid {
 	* Calculates hamming window coefficients.
 	*/
 	template <typename T>
-	void Blackman(int N, T amp, vector<T> &win) {
+	void Blackman(uint64_t N, T amp, vector<T> &win) {
 		for (int i = 0; i<(N + 1) / 2; ++i) {
 			win[i] = amp * T(0.42 - 0.50*cos(M_PI * 2 * i / (N - 1.0))
 				+ 0.08*cos(2 * M_PI * 2 * i / (N - 1.0)));
@@ -95,7 +96,7 @@ namespace vocaloid {
 	* Calculates kasier window coefficients.
 	*/
 	template <typename T>
-	void Kaiser(int N, float beta, vector<T> &win) {
+	void Kaiser(uint64_t N, float beta, vector<T> &win) {
 		win.reserve(N + 1);
 		for (int i = 0; i <= N; i++) {
 			win[i] = BesselI0(beta * sqrt(1.0 - pow((i - N * 0.5) / (N * 0.5), 2.0))) / BesselI0(beta);
@@ -104,7 +105,7 @@ namespace vocaloid {
 
 
 	template<typename T>
-	void GenerateWin(WINDOW_TYPE type, int length, vector<T> &output, float extra = 1.0) {
+	void GenerateWin(WINDOW_TYPE type, uint64_t length, vector<T> &output, float extra = 1.0) {
 		switch (type) {
 		case WINDOW_TYPE::HANNING:
 			Hanning(length, output);
